+++ internal/infrastructure/database/migrations.go
@@ -1,8 +1,1 @@
 package database
-
-import (
-	_ "embed"
-)
-
-//go:embed migrations.sql
-var migrationsSQL string